requirements.txt
opencv-python==4.8.1.78
numpy==1.26.2
face-recognition==1.3.0
Pillow==10.1.0
dlib==19.24.2 